removing syntax highlighting
The current implementation of syntax highlighting was only ever intended
as an initial proof of concept (largely due to performance reasons).
Finding a proper solution to this and supporting more than just Rust,
shell and the plumbing rules file types is going to be something that
requires proper thought and likely multiple changes to the internal data
structures.

I've considered using treesitter but I'm not overly happy with having to
track each edit twice (once for the buffer content and once for tree
sitter) so I'll need to see if there is a way to handle this without
baking in too many assumptions about the structure of buffer contents.
Given that a key part of ad is the assertion that all text inside of ad
is treated equally, I'm not a massive fan of having to set up different
syntax support for custom buffers and extension programs that users
write (or supporting rich markup inside of ad buffers).

If possible it would be nice to push this out to only live on the UI
side of things, and to support running ad in a headless mode capable of
driving a UI efficiently so that UI implementers were free to handle
that side of things in the way they want.