fix(ninep/client): correct terminology around namespaces on unix
Since this is unix specific, I'm basing this on 9p(1), namespace(1) and
intro(4) from plan9port.

- The namespace is the /tmp/ns.$USER.$DISPLAY directory that consists
  of unix sockets to various 9p servers (without any nesting).
- This can be used by e.g. 9p(1) to create the illusion of a coherent
  namespace where you can 9p read foo/bar which actually dispatches
  to the singular /tmp/ns.$USER.$DISPLAY/foo socket.

I've called the first path component, which always needs to be given
`server_name`. I don't know if it has a more specific term in plan9port,
though. My main aim is to prevent a confusion with namespace which (in
plan9port) refers to the directory that contains the 9p unix sockets.