refactor(ninep): use PathBuf for unix::namespace()
The use of String for most things 9p is a good fit since its messages
use UTF-8 encoding. The namespace is a POSIX path, though, and may
theoretically be an arbitrary byte sequence, so it's better to use
PathBuf (and OsString). Doing this consistently in ninep should also
create a nice distinction between 9p paths and POSIX paths via their
types.

Additionally, we no longer need to duplicate logic between socket_path()
and unix_socket() in server because we can use PathBuf::parent().